Reduce size of entry
There are a lot of information in the struct '__PHONE_OTHER_ENTRT'.
But actually two functions ,find() and append(), just need to use
the 'lastName'. The size of struct now is too big to cause many
cache-miss. Because of that,I try to comment the redundant code and
reduse the size of struct in phonebook_opt.h. The result shows that
reducing the size of struct can reduse the cache-miss rate. Next time,
I want to reduce the size of struct and keep these detail information
at the same time.