Reduce size of entry
There are a lot of entry in the struct '__PHONE_OTHER_ENTRT'.
But actually two functions ,find() and append(), just need to use
the 'lastName'. The size of struct now is too big to cause many
cache-miss. Because of that,I try to comment the redudant entry and
reduce the size of struct in phonebook_opt.h. The result shows that
reducing the size of struct can reduce the cache-miss rate. Next time,
I want to reduce the size of struct and keep these detail entry at the
same time.