Implement hash table
In order to make the 'findName()' faster and reduce the cache-miss
rate. I try to implement BKDR hash table and observe the result.
It shows that using hash table can let the 'findName()' faster.
But it make the 'append()' become slower at the same time. In other
hand, the cache-miss rate has been decreased in this version.

I will try to implement other hash methods and compare their efficiency.