Add BST tree to accelerate findName and append
In this version, I try to add BST tree to make the 'findName()'
and 'append()' become faster. Because of the BST tree's worst
case in search tree node is O(n) and the average time complexity
is O(logn), I think it is a good way to accelerate 'findName'
and 'append()'. And the result shows that the 'findName()'
become faster than last version. The 'append()' is faster
than the version with hash table, too. But the total time is
increased. The reason is that BST tree's building time is not
a small number. Although the time of 'findName' and 'append()'
is decreased but we need to plus the time of building BST tree,
so the total time cost is increased.